Fixed aspect ratio, with floats in UpdateProjectionMatrix function

+++ DirectX11Engine/Camera.cpp
@@ -21,7 +21,7 @@ void Camera::UpdateViewMatrix()
 	ViewMatrix = XMMatrixLookToLH(Position, ForwardVector, UpVector);
 }
 
-void Camera::UpdateProjectionMatrix(int Width, int Height)
+void Camera::UpdateProjectionMatrix(float Width, float Height)
 {
 	ProjectionMatrix = XMMatrixPerspectiveFovLH(Math::DegreesToRadian(80.0f), Width / Height, 0.1f, 10000.0f);
 }
